% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytoxcalc-package.R
\docType{package}
\name{shinytoxcalc-package}
\alias{shinytoxcalc}
\alias{shinytoxcalc-package}
\title{shinytoxcalc: A Graphical Interface to the toxcalc EPA WET Methods}
\description{
A Shiny application providing a point-and-click interface to the 'toxcalc' package, which implements the US EPA Whole Effluent Toxicity statistical methods. Data are entered by pasting, by uploading a comma-separated file, or by uploading a spreadsheet laid out to the supplied template. Results are presented as the decision trail that justifies the selected test, alongside the endpoints and a concentration-response plot, and can be exported as a report, a spreadsheet, or the R code that reproduces the analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/beckyfisher/shinytoxcalc}
  \item Report bugs at \url{https://github.com/beckyfisher/shinytoxcalc/issues}
}

}
\author{
\strong{Maintainer}: Rebecca Fisher \email{r.fisher@aims.gov.au}

Authors:
\itemize{
  \item Rebecca Fisher \email{r.fisher@aims.gov.au}
}

}
\keyword{internal}
